+++ main/lucid/xpr/Registry.h
@@ -18,22 +18,22 @@ public:
 
 	virtual ~Registry() = default;
 
-	size_t operator()(std::string const &symbol) const
+	uint64_t operator()(std::string const &symbol) const
 	{
 		return index_of(symbol);
 	}
 
-	size_t add(std::string const &symbol, float64_t value = 0.0)
+	uint64_t add(std::string const &symbol, float64_t value = 0.0)
 	{
 		auto iter = indices.find(symbol);
 		if (iter != indices.end())
 		{
-			size_t index = iter->second;
+			uint64_t index = iter->second;
 			set(index, value);
 			return index;
 		}
 
-		size_t index = entries.size();
+		uint64_t index = uint64_t(entries.size());
 
 		entries.push_back(value);
 		indices[symbol] = index;
@@ -42,9 +42,9 @@ public:
 		return index;
 	}
 
-	void set(size_t index, float64_t value)
+	void set(uint64_t index, float64_t value)
 	{
-		entries[index] = value;
+		entries[size_t(index)] = value;
 	}
 
 	void set(std::string const &symbol, float64_t value)
@@ -52,7 +52,7 @@ public:
 		set(index_of(symbol), value);
 	}
 
-	size_t index_of(std::string const &symbol) const
+	uint64_t index_of(std::string const &symbol) const
 	{
 		auto const iter = indices.find(symbol);
 		if (iter == indices.end())
@@ -60,7 +60,7 @@ public:
 		return iter->second;
 	}
 
-	float64_t value_of(size_t index) const
+	float64_t value_of(uint64_t index) const
 	{
 		return entries[index];
 	}
@@ -70,7 +70,7 @@ public:
 		return value_of(index_of(symbol));
 	}
 
-	std::string const &symbol_for(size_t index) const
+	std::string const &symbol_for(uint64_t index) const
 	{
 		auto const iter = symbols.find(index);
 		if (iter == symbols.end())
@@ -82,8 +82,8 @@ private:
 	std::string const undefined = "<undef>";
 
 	std::vector<float64_t> entries;
-	std::unordered_map<std::string, size_t> indices;
-	std::unordered_map<size_t, std::string> symbols;
+	std::unordered_map<std::string, uint64_t> indices;
+	std::unordered_map<uint64_t, std::string> symbols;
 };
 
 LUCID_XPR_END
